refactor: clear views

+++ resources/views/admin/users/page.blade.php
@@ -10,40 +10,41 @@
 @else
     @foreach ($users as $user)
         <x-admin.table.row>
-            <x-admin.table.cell
-                :font_light="false"
-                class="w-1/2"
-            >
+            <x-admin.table.cell :font_light="false">
                 {{ $user->name }}
             </x-admin.table.cell>
-            <x-admin.table.cell class="w-1/2">
+            <x-admin.table.cell>
                 {{ $user->email }}
             </x-admin.table.cell>
-            <x-admin.table.cell>
-                <x-admin.page-link
-                    :params="$user"
-                    route="admin.users.edit"
-                    title="Edit"
-                >
-                    <div class="rounded p-1 hover:bg-indigo-700 hover:text-white">
-                        <x-heroicon-o-pencil-square class="h-4" />
-                    </div>
-                </x-admin.page-link>
-            </x-admin.table.cell>
-            <x-admin.table.cell>
-                <button
-                    class="rounded p-1 hover:bg-indigo-700 hover:text-white"
-                    hx-confirm="{{ trans('Are you sure?') }}"
-                    hx-delete="{{ route('admin.users.destroy', $user) }}"
-                    hx-disabled-elt="this"
-                    hx-target="closest tr"
-                    title="{{ trans('Remove') }}"
-                    type="button"
-                >
-                    <x-heroicon-o-trash class="content h-4" />
-                    <x-admin.loading class="h-4 w-4" />
-                </button>
-            </x-admin.table.cell>
+            @can('admin.users.edit')
+                <x-admin.table.cell>
+                    <x-admin.page-link
+                        :params="$user"
+                        route="admin.users.edit"
+                        title="Edit"
+                    >
+                        <div class="rounded p-1 hover:bg-indigo-700 hover:text-white">
+                            <x-heroicon-o-pencil-square class="h-4" />
+                        </div>
+                    </x-admin.page-link>
+                </x-admin.table.cell>
+            @endcan
+            @can('admin.users.destroy')
+                <x-admin.table.cell>
+                    <button
+                        class="rounded p-1 hover:bg-indigo-700 hover:text-white"
+                        hx-confirm="{{ trans('Are you sure?') }}"
+                        hx-delete="{{ route('admin.users.destroy', $user) }}"
+                        hx-disabled-elt="this"
+                        hx-target="closest tr"
+                        title="{{ trans('Remove') }}"
+                        type="button"
+                    >
+                        <x-heroicon-o-trash class="content h-4" />
+                        <x-admin.loading class="h-4 w-4" />
+                    </button>
+                </x-admin.table.cell>
+            @endcan
         </x-admin.table.row>
     @endforeach
     <x-admin.table.row>
@@ -51,9 +52,9 @@
             :font_light="false"
             colspan="all"
         >
-            <div class="min-h-8 text-slate-500/80">
-                <div class="content flex w-full">
-                    @if ($users->hasMorePages())
+            @if ($users->hasMorePages())
+                <div class="min-h-8 text-slate-500/80">
+                    <div class="content flex w-full">
                         <button
                             class="m-auto h-min rounded border border-slate-200 px-6 py-1.5 text-xs uppercase text-slate-700 hover:bg-neutral-500 hover:bg-opacity-10"
                             hx-disabled-elt="this"
@@ -65,10 +66,10 @@
                         >
                             @lang('Load more')
                         </button>
-                    @endif
+                    </div>
+                    <x-admin.loading class="h-5" />
                 </div>
-                <x-admin.loading class="h-5" />
-            </div>
+            @endif
         </x-admin.table.cell>
     </x-admin.table.row>
 @endempty
